// apps/web/src/api/generated.ts
export interface components {
    schemas: {
// ...
        AuthTokens: {
            accessToken: string;
            refreshToken: string;
            user: {
                id: number;
                /** @enum {string} */
                role: "admin" | "gestor" | "financeiro" | "motorista";
            };
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}